# [Performance] Use new Azure custom images