<!-- Describe the changes in this PR -->
Use the new custom image for Azure's VM creation. 

Example performance change (with fixed region):
|  VM Type 💻   | Old Provision 🕐                 | New Provision 🕐     | % Speedup ✅ |
| --------------- | ---------------------------- | ----------- | ----------- |
| Azure GPU      |   4min    | 2min 40s   | 35% (1.5x) |
| Azure CPU       |  4min  | 2min |  50% (2x) |

Example benchmark commands:
CPU: `time sky launch --cloud=azure --region=centralus -y`
GPU: `time sky launch --cloud=azure --region=centralus -y --gpus=A100-80GB:1`

<!-- Describe the tests ran -->
<!-- Unit tests (tests/test_*.py) are part of GitHub CI; below are tests that launch on the cloud. -->

Tested (run the relevant ones):

- [x] Code formatting: `bash format.sh`
- [x] Relevant individual smoke tests: 
`pytest tests/test_smoke.py::test_minimal --azure`
`pytest tests/test_smoke.py::test_huggingface --azure`
`pytest tests/test_smoke.py::test_job_queue_with_docker --azure`
`pytest tests/test_smoke.py::test_cancel_azure`